better handling of nulls

+++ persistence/src/main/java/org/smallmind/persistence/orm/jdo/JDODao.java
@@ -1,28 +1,28 @@
 /*
  * Copyright (c) 2007, 2008, 2009, 2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017 David Berkman
- * 
+ *
  * This file is part of the SmallMind Code Project.
- * 
+ *
  * The SmallMind Code Project is free software, you can redistribute
  * it and/or modify it under either, at your discretion...
- * 
+ *
  * 1) The terms of GNU Affero General Public License as published by the
  * Free Software Foundation, either version 3 of the License, or (at
  * your option) any later version.
- * 
+ *
  * ...or...
- * 
+ *
  * 2) The terms of the Apache License, Version 2.0.
- * 
+ *
  * The SmallMind Code Project is distributed in the hope that it will
  * be useful, but WITHOUT ANY WARRANTY; without even the implied warranty
  * of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
  * General Public License or Apache License for more details.
- * 
+ *
  * You should have received a copy of the GNU Affero General Public License
  * and the Apache License along with the SmallMind Code Project. If not, see
  * <http://www.gnu.org/licenses/> or <http://www.apache.org/licenses/LICENSE-2.0>.
- * 
+ *
  * Additional permission under the GNU Affero GPL version 3 section 7
  * ------------------------------------------------------------------
  * If you modify this Program, or any covered work, by linking or
@@ -60,24 +60,26 @@ public abstract class JDODao<I extends Serializable & Comparable<I>, D extends D
 
   public D get (Class<D> durableClass, I id) {
 
-    VectoredDao<I, D> vectoredDao;
-    D durable;
+    if (id != null) {
 
-    if ((vectoredDao = getVectoredDao()) == null) {
-      if ((durable = acquire(durableClass, id)) != null) {
+      VectoredDao<I, D> vectoredDao;
+      D durable;
 
-        return durable;
-      }
-    }
-    else {
-      if ((durable = vectoredDao.get(durableClass, id)) != null) {
+      if ((vectoredDao = getVectoredDao()) == null) {
+        if ((durable = acquire(durableClass, id)) != null) {
 
-        return durable;
-      }
+          return durable;
+        }
+      } else {
+        if ((durable = vectoredDao.get(durableClass, id)) != null) {
 
-      if ((durable = acquire(durableClass, id)) != null) {
+          return durable;
+        }
 
-        return vectoredDao.persist(durableClass, durable, UpdateMode.SOFT);
+        if ((durable = acquire(durableClass, id)) != null) {
+
+          return vectoredDao.persist(durableClass, durable, UpdateMode.SOFT);
+        }
       }
     }
 
@@ -87,7 +89,7 @@ public abstract class JDODao<I extends Serializable & Comparable<I>, D extends D
   @Override
   public D acquire (Class<D> durableClass, I id) {
 
-    return durableClass.cast(getSession().getNativeSession().getObjectById(durableClass, id));
+    return (id == null) ? null : durableClass.cast(getSession().getNativeSession().getObjectById(durableClass, id));
   }
 
   public List<D> list () {
@@ -109,34 +111,42 @@ public abstract class JDODao<I extends Serializable & Comparable<I>, D extends D
     return new IterableIterator<D>(getSession().getNativeSession().getExtent(getManagedClass()).iterator());
   }
 
-  public D detach (D object) {
+  public D detach (D durable) {
 
-    return getManagedClass().cast(getSession().getNativeSession().detachCopy(object));
+    return (durable == null) ? null : getManagedClass().cast(getSession().getNativeSession().detachCopy(durable));
   }
 
   public D persist (Class<D> durableClass, D durable) {
 
-    D persistentDurable;
-    VectoredDao<I, D> vectoredDao = getVectoredDao();
+    if (durable != null) {
 
-    persistentDurable = durableClass.cast(getSession().getNativeSession().makePersistent(durable));
+      D persistentDurable;
+      VectoredDao<I, D> vectoredDao = getVectoredDao();
 
-    if (vectoredDao != null) {
+      persistentDurable = durableClass.cast(getSession().getNativeSession().makePersistent(durable));
 
-      return vectoredDao.persist(durableClass, persistentDurable, UpdateMode.HARD);
+      if (vectoredDao != null) {
+
+        return vectoredDao.persist(durableClass, persistentDurable, UpdateMode.HARD);
+      }
+
+      return persistentDurable;
     }
 
-    return persistentDurable;
+    return null;
   }
 
   public void delete (Class<D> durableClass, D durable) {
 
-    VectoredDao<I, D> vectoredDao = getVectoredDao();
+    if (durable != null) {
 
-    getSession().getNativeSession().deletePersistent(durable);
+      VectoredDao<I, D> vectoredDao = getVectoredDao();
 
-    if (vectoredDao != null) {
-      vectoredDao.delete(durableClass, durable);
+      getSession().getNativeSession().deletePersistent(durable);
+
+      if (vectoredDao != null) {
+        vectoredDao.delete(durableClass, durable);
+      }
     }
   }
